requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
torch>=1.11.0
torchvision>=0.12.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.62.0
PyYAML>=6.0
scikit-learn>=1.0.0